Bugfixes
Added winner indicator.

Fixed ties not working as expected. Still an issue with _pot division and a bug in the tie detection method that needs to be fully debugged.

Next on the list is to add the blind and fix the raise methods after remaining bugs in the tie calculation are resolved.