Minor Update
Added blind betting.

Player's are now taking thier turns properly.

Added personalities to the player class.

Fixed an issue with the tie buffers.

Next step is to fix a bug with the raise button and allow the AI players to do more than hold or match each round. WIll not be implementing TexasOdds until after that framework since the former is a necessity for a satisfying game.