Minor Update
UI updates.

New fix for raise system is being worked on. Must decide wether to continue as is, or rework the system so that the state machine returns to the mainwindow class at the end of each ai turn. Might be necessary to add in any sort of animation but will set the application back another week.