Added case for when AI player leaves game.
Updates for GUI.

Found issue in fold, players seem to continue to draw although the deck is set to null, need to look into order of methods and stop this.

Game sometimes stalls and only pressing fold ends, but results in a fold victory, need to look into this when time permits.

Tie checker produced a rare error, have added debugging code, but need to wait for the error to reoccur in order to debug it.

Made a change to circular int which will not harm its use in this program but will make it un-reusable code due to inconsistency. I don't want to program it as well as I could since it is just a waste of time.